feat(fuzzy): use `token_sort_ratio` instead of `token_set_ratio` and sort by ratio
the problem with using token_set_ratio:
will return `100` for all queries which contain only one of the words in the given title
searching for `the` would return `100` for every title with `the` in it
originally this was simply used to be able to shorten search terms (e.g. `/watch Pacific`)
this is obviously problematic, since we can't narrow down the fuzzy search to a single result
(which should almost always be possible with a few exceptions where the client has to do this
by year)
I'm not sure whether a year parameter should be added here, this wouldn't really be "fuzzy"
anymore. This should probably be used by callers of this function